requirements.txt
llama-cpp-python
sentence-transformers
faiss-cpu
langchain-docling
docling-parse
huggingface-hub
